move and simplify test
